requirements.txt
pyinstaller
pandas
matplotlib
openpyxl
xlrd
chardet
psutil
python-dateutil
pycryptodome